Update gps2trips.R
started the "getActivities" function. So far, it takes in the cumulative distance table and adds two columns onto it. One column is the change in cumulative distance for every row, and the other column is the moving average of the first column. They are both "working" in the sense that they are not getting errors and are mutating correctly. However, the moving average column is always getting 0 when that should not be the case. I need to double check that the rolling average function is correct and see if there is a way to allow decimal results instead of always numeric. Either one of those could be the problem.