Added getClusters function
getClusters further filters the cleaned data by just examining one date from the respondent. It then adds on a time,x, and y column. Using the dbscan library, it then calculates the number of clusters in that given day from thresholds "eps" and "minPoints". As of right now, a radius (eps) of 50 feet and a minimum number of 20 points is used to be considered a cluster. These thresholds give the lowest and most accurate number of clusters as of right now. Still need to figure out a way to include repeating clusters (for example, multiple trips back home). 

Also, I am still messing around with the targets for this function. For some reason, when I create a target with this function the way it is, the target is NULL. Therefore, I am committing the code as it is right now before I start splitting this function up and trying new ways to get the target to work. I would like a numClusters target and a separate visual_clusters target which is the color-coded map showing the different clusters.
 
Overall, this code works and gives a relatively probable number for the number of trips made using the dbscan library. I am going to mess around with the targets now and try to get more accurate numbers of clusters. Just come back to this code if I end up messing everything up by doing all of that :)