gpsactivs data extraction and functions with caps data
This new gps2trips script essentially scraps a lot of the density and time-based algorithm functions. It keeps the cleaning aspects as well as transforming latitude and longitudes into geometrical coordinates. This code was written by Dr. Macfarlane with some sample trajectory data, but I altered it to point to our CAPS data. It works with our caps data. I added the "cluster_info" object to the global environment. "cluster_info" uses the dbscan_te() function from the gpsactivs package written by Dr. Macfarlane. The next step I have to do is mess around with the parameters of that function to see if we can get reasonable outcomes for several different trace days.