Add dbscan::kNNdistplot
Figured out why the trajectory plot was not working correctly as noted above the leaflet code. Then, the dbscan::kNNdistplot was added and can be used to draw the k distance plot given a data matrix and k value. The k value is the same thing as the minPoints. Where the curvature changes in the k distance plot is where the optimal eps value is given a certain minPoints parameter. Right now, it is saying that caps_tr is not numeric and I am trying to fix it so that is is, so that the kNNdistplot function will work.